still was not fucking ready, should be now

+++ src-tauri/src/main.rs
@@ -13,12 +13,12 @@ use serde_json;
 fn main() {
     first_install();
     tauri::Builder::default()
-        .invoke_handler(tauri::generate_handler![get_links_json, add_link])
+        .invoke_handler(tauri::generate_handler![get_links_json, add_link, delete_link])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
 struct Link {
     name: String,
     url: String
@@ -27,13 +27,32 @@ struct Link {
 // add links to file to remember
 #[tauri::command]
 fn add_link(name: String, url: String) {
-    let mut links = get_links();
-    let link: Link = Link { name: name, url: url };
-    links.push(link);
-    let json = serde_json::to_string(&links).unwrap();
-    let mut data_file: PathBuf = dirs::home_dir().unwrap();
-    data_file.push(".siponen/links.json");
-    fs::write(data_file, json).unwrap();
+    if !link_exists(name.clone()) {
+        let mut links = get_links();
+        let link: Link = Link { name: name, url: url };
+        links.push(link.clone());
+        let json = serde_json::to_string(&links).unwrap();
+        let mut data_file: PathBuf = dirs::home_dir().unwrap();
+        data_file.push(".siponen/links.json");
+        fs::write(data_file, json).unwrap();
+        println!("Added link to file: {link:?}")
+    }
+}
+
+// delete link with it's name
+#[tauri::command]
+fn delete_link(name: String) {
+    if link_exists(name.clone()) {
+        let mut links = get_links();
+        links.retain(|x| x.name != name);
+
+        let json = serde_json::to_string(&links).unwrap();
+        let mut data_file: PathBuf = dirs::home_dir().unwrap();
+        data_file.push(".siponen/links.json");
+        fs::write(data_file, json).unwrap();
+
+        println!("Deleted link: {name:?}");
+    }
 }
 
 // gets links so rust can understand them
@@ -42,6 +61,7 @@ fn get_links() -> Vec<Link> {
     data_file.push(".siponen/links.json");
     let contents = fs::read_to_string(data_file).unwrap();
     let links: Vec<Link> = serde_json::from_str(&contents).unwrap();
+    println!("Got links from file");
     return links
 }
 
@@ -52,7 +72,21 @@ fn get_links_json() -> serde_json::Value {
     data_file.push(".siponen/links.json");
     let contents = fs::read_to_string(data_file).unwrap();
     let links: Vec<Link> = serde_json::from_str(&contents).unwrap();
+    println!("Got links from file for frontend");
     return serde_json::json!(&links)
+}
+
+// check if name already exists
+fn link_exists(name: String) -> bool {
+    let links = get_links();
+    for i in links {
+        if i.name == name {
+            println!("link {name:?} exists");
+            return true
+        }
+    }
+    println!("link {name:?} does not exist");
+    return false
 }
 
 // If program was installed first time this function creates the data directory and data file for links
@@ -72,6 +106,4 @@ fn first_install() {
         fs::write(file, json).unwrap();
         println!("Links data file created\n");
     }
-    std::thread::sleep(std::time::Duration::from_secs(2));
-    add_link("kakka".to_string(), "riveria.fi".to_string());
 }
